fix(search): address #372 confirmed review findings
Fixes all medium/high/low confirmed findings from the wave-4 #372
review pass on the lexical AND-intersect branch:

[high] Double-offset on text + temporal-sort path: pass `offset: None`
into QueryConfig when `flags.temporal_sort.is_some()` so the pre-sort
verify step never consumes the offset; the post-sort drain in run_query
is the single correct application. Add `test_double_offset_guard_*`.

[medium] Dead code — delete `intersect_posting_doc_ids` (reader.rs) and
its `#[allow(dead_code)]` annotation. The live path (`search_exact_intersection`)
inlines the identical ~45-line merge; removing the dead copy eliminates
the two-copy drift risk (quality.md: "Delete dead code").

[medium] `offset: Option<u64>` → `Option<usize>` in `Flags`, `QueryConfig`,
and `parse_offset_value` to match `limit: usize` and `SearchQuery::offset:
Option<usize>`. Removes four `as usize` casts at consumption sites.

[medium] Split `test_ac15a_short_query_fallback_5000_files_sla` into:
- `test_ac15a_short_query_fallback_5000_files_recall` — recall + high-id
  assertions, always-on (both debug and release), guards AD-372-4 in every
  `cargo test` run.
- `test_ac15a_short_query_fallback_5000_files_sla` — wall-clock SLA only,
  `#[cfg(not(debug_assertions))]`. Doc comment previously claimed the recall
  assertions ran under debug; that was false.

[low] Stale `intersect_posting_doc_ids →` reference in AC#5 error-chain
doc comment in reader_tests.rs: update to reflect the actual shipping path
(`lookup_postings → decode_postings_varint → search_exact_intersection →
search()`).

[low] Update `resolve_paths_and_snippets_verified` fan-out bound doc
comment: add the exact-symbol sub-path (bounded by intersection size, not
K×limit) alongside the existing multi-word and blast-radius entries.

[low] Strengthen `has_high_id_file` assertion: replace the brittle
path-substring OR-chain with a direct `r.path.contains("f4999.rs")`
check on a known specific high-id file.

[medium/PF-008] Update KNOWLEDGE.md score-field semantics: single-token
queries now score by raw occurrence count (AD-372-6, length-norm-free),
not BM25F magnitude. Multi-word UNION path retains BM25F semantics.

+++ crates/rskim/src/cmd/search/query.rs
@@ -284,7 +284,7 @@ pub(super) fn execute_query_with_manifest(
     // verify-then-truncate-LAST invariant.  Offset is applied HERE (post-verify)
     // on BOTH the exact-symbol and multi-word paths so that page boundaries are
     // consistent regardless of how many pre-verify candidates are dropped.
-    let effective_offset = config.offset.unwrap_or(0) as usize;
+    let effective_offset = config.offset.unwrap_or(0);
     let results = resolve_paths_and_snippets_verified(
         &raw_results,
         &sorted,
@@ -457,7 +457,7 @@ fn run_compound_query(
     // `skim search "foo" --ast try-catch --offset 10` paginates correctly.
     // The RRF recomposition does NOT apply offset; pagination is handled here,
     // post-verify, as on the pure-lexical path.
-    let effective_offset = config.offset.unwrap_or(0) as usize;
+    let effective_offset = config.offset.unwrap_or(0);
     let results = resolve_paths_and_snippets_verified(
         &recomposed,
         ctx.sorted,
@@ -594,7 +594,7 @@ fn run_blast_radius_composite_query(
     // `skim search "foo" --blast-radius src/x.rs --offset 10` paginates correctly.
     // Applied post-verify (`.skip` before `.take`), consistent with the pure-lexical
     // and compound paths.
-    let effective_offset = config.offset.unwrap_or(0) as usize;
+    let effective_offset = config.offset.unwrap_or(0);
     let results: Vec<super::types::ResolvedResult> = ranked
         .iter()
         .filter_map(|&(fid, composite_score)| {
@@ -728,9 +728,12 @@ struct SnippetVerifyParams<'a> {
 ///
 /// The worst-case file-read count equals `raw_results.len()`, which is bounded
 /// by the caller's `sq.limit` — itself bounded to:
-///   - Pure-lexical path:   `max(limit × LEXICAL_CANDIDATE_POOL_K, 100)` = `max(5N, 100)`.
-///   - Blast-radius path:   `max(limit × BLAST_CANDIDATE_POOL_K, 100)`  = `max(10N, 100)`.
-///   - Compound text+AST:   `|ast_set|` (exact AST match count; no K multiplier, AD-356-1).
+///   - Pure-lexical exact-symbol path: `|intersection|` (AND of query trigram posting
+///     lists; `sq.limit = None` so the full intersection reaches verify). Bounded by
+///     posting list sizes, not corpus size (AD-372-2 superset invariant).
+///   - Pure-lexical multi-word path:   `max(limit × LEXICAL_CANDIDATE_POOL_K, 100)` = `max(5N, 100)`.
+///   - Blast-radius path:              `max(limit × BLAST_CANDIDATE_POOL_K, 100)`  = `max(10N, 100)`.
+///   - Compound text+AST:              `|ast_set|` (exact AST match count; no K multiplier, AD-356-1).
 ///
 /// The fan-out is therefore O(K × limit) file reads — bounded for any fixed K and
 /// user-supplied `--limit`.  Calibrating K for large corpora is tracked in #361.
